refactor: move logic back to plugin
